Reorganise scripts
